Fix regression in lifted semigroup sum. Move lifted definition in Monoid.
